move all structs to structs file
